Switch trie backend API to not return itself
With the redis client, all changes happen in an actor - no local variable is mutated.
Of course that doesn't work with the Memory backend which uses a Dict. Thus the Memory backend is disabled for now. It needs to be re-implemented with an actor that maintains the Dict which just receives messages via the process subject.